Fix race conditions during stderr redirect.
I had intended to do this before, but forgot about it.  Without this, it
is possible to bad stderr input to show up when RustBox races to
initialize against other threads trying to do so (it's not memory
unsafe, just causes exception spam).  I'm not sure whether anyone will
actually try to do this, given that RustBox doesn't work on multiple
threads, but it's still better to be correct here :)