Added keyboard module
rustbox::poll_event & rustbox::peek_event now take a boolean parameter
which determines the type of event you want to get back from the method.

Setting this parameter to true will return the Raw variants of the
different events, which contain the raw event data. For instance,
KeyEventRaw which will contain the integers representing the key's ch,
mod and key values.

Instead, setting this param to false will return the processed variant
of the events. Using the above example, the result would be KeyEvent
which would contain a variant of rustbox::Key.

To upgrade change any uses of rustbox::poll_event & rustbox::peek_event
to take a boolean parameter. To update to use rustbox::Key, set this
param to false, and use rustbox::Key. To keep current behaviour, set
this param to true.