make RustBox type Send, but still not Sync
The RustBox type is not thread safe, in the sense that multiple
threads may never alter the mutable state of a shared RustBox
at the same time, so it should not be marked Sync.
However, a RustBox type *can* safely be sent between threads so
it *should* be marked Send.

This commit makes it possible to wrap a RustBox struct inside
a Mutex type, so that it can still be used when the programmer
decides to do the syncronization theirself.